+++ lib/rubykassa/signature_generator.rb
@@ -2,21 +2,25 @@
 module Rubykassa
   module SignatureGenerator
     def generate_signature_for kind
-      raise ArgumentError, "Available kinds are only :payment, :result or :success" unless [:success, :payment, :result].include? kind
-      custom_params = @params.keys.select {|key| key =~ /^shp/ }.sort.map {|key| "#{key}=#{params[key]}"}
-      custom_params_string = custom_params.present? ? ":#{custom_params}" : ""
-      
-      Digest::MD5.hexdigest(params_string(kind, custom_params_string))
+      raise ArgumentError, "Available kinds are only :payment, :result or :success" unless [:success, :payment, :result].include? kind      
+      Digest::MD5.hexdigest(params_string(kind))
     end
 
-    def params_string kind, custom_params_string
+    def params_string kind
       string = case kind
       when :payment
-        [Rubykassa.login, @total, @invoice_id, Rubykassa.first_password].join(":") + custom_params_string
+        [Rubykassa.login, @total, @invoice_id, Rubykassa.first_password, custom_params].flatten.join(":")
       when :result
-        [@total, @invoice_id, Rubykassa.second_password].join(":") + custom_params_string
+        [@total, @invoice_id, Rubykassa.second_password, custom_params].flatten.join(":")
       when :success
-        [@total, @invoice_id, Rubykassa.first_password].join(":") + custom_params_string
+        [@total, @invoice_id, Rubykassa.first_password, custom_params].flatten.join(":")
+      end
+    end
+
+    def custom_params
+      @params.sort.inject([]) do |result, element|
+        result << element.join("=") if element[0] =~ /^shp/
+        result
       end
     end
   end